+++ src/main/java/de/samply/frontend/dto/ProjectBridgehead.java
@@ -1,10 +1,12 @@
 package de.samply.frontend.dto;
 
+import com.fasterxml.jackson.annotation.JsonInclude;
 import de.samply.project.state.ProjectBridgeheadState;
 import de.samply.query.QueryState;
 
 import java.time.Instant;
 
+@JsonInclude(JsonInclude.Include.NON_NULL)
 public record ProjectBridgehead(
         String projectCode,
         String bridgehead,