+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.0.1 - 2024-11-27]
+## [0.0.1 - 2024-11-28]
 ### Added
 - First version of the project
 - Spring Application
@@ -159,3 +159,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Integration of action explanatios in fetch Actions
 - Action explanations templates
 - Email Context
+- Extend email context
+- Thymeleaf Dialect for email context variables